Move add list tag function
